server/ets/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
blis==0.7.7
bs4==0.0.1
catalogue==2.0.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cymem==2.0.6
Django==4.0.3
djangorestframework==3.13.1
djutil==0.2
djutils==1.0.15
en-core-web-sm==3.3.0
idna==3.3
Jinja2==3.1.2
joblib==1.1.0
langcodes==3.3.0
MarkupSafe==2.1.1
murmurhash==1.0.7
nltk==3.7
numpy==1.22.4
packaging==21.3
pandas==1.4.2
pathy==0.6.1
Pillow==9.0.1
preshed==3.0.6
psycopg2-binary==2.9.3
pydantic==1.8.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2021.3
regex==2022.4.24
requests==2.27.1
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
sqlparse==0.4.2
srsly==2.4.3
thinc==8.0.16
threadpoolctl==3.1.0
tqdm==4.64.0
typer==0.4.1
typing-extensions==4.2.0
urllib3==1.26.9
wasabi==0.9.1
